filess/requirements.txt
asttokens==3.0.1
audioread==3.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.25.2
fsspec==2026.2.0
idna==3.11
ipython==8.38.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.39.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
numba==0.56.4
numpy==1.23.5
opencv-python==4.13.0.92
packaging==26.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
pillow_heif==1.3.0
platformdirs==4.9.4
pooch==1.9.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sounddevice==0.5.5
soundfile==0.13.1
soxr==1.0.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
xgboost==1.7.6
